+++ tests/TestCase/Driver/MysqlTest.php
@@ -169,7 +169,7 @@ class MysqlTest extends DriverTestCase
 
     /**
      * Test for `export()` method on failure
-     * @expectedException Cake\Network\Exception\InternalErrorException
+     * @expectedException RuntimeException
      * @expectedExceptionMessage Failed with exit code `2`
      * @test
      */
@@ -184,7 +184,7 @@ class MysqlTest extends DriverTestCase
 
     /**
      * Test for `import()` method on failure
-     * @expectedException Cake\Network\Exception\InternalErrorException
+     * @expectedException RuntimeException
      * @expectedExceptionMessage Failed with exit code `1`
      * @test
      */