+++ tests/TestCase/Driver/SqliteTest.php
@@ -79,7 +79,7 @@ class SqliteTest extends DriverTestCase
 
     /**
      * Test for `export()` method on failure
-     * @expectedException Cake\Network\Exception\InternalErrorException
+     * @expectedException RuntimeException
      * @expectedExceptionMessage Failed with exit code `1`
      * @test
      */
@@ -113,7 +113,7 @@ class SqliteTest extends DriverTestCase
 
     /**
      * Test for `import()` method on failure
-     * @expectedException Cake\Network\Exception\InternalErrorException
+     * @expectedException RuntimeException
      * @expectedExceptionMessage Failed with exit code `1`
      * @test
      */