+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # 2.x branch
 ## 2.4 branch
 ### 2.4.0
+* fixed bug trying to email a nonexistent backup;
 * `VALID_COMPRESSIONS` and `VALID_EXTENSIONS` constants have been replaced by
     `getValidCompressions()` and `getValidExtensions()` methods provided by the
-    `BackupTrait` class.
+    `BackupTrait` class;
+* replaced `InternalErrorException` with `InvalidArgumentException` and 
+    `RuntimeException`. This allows compatibility with CakePHP 3.6 branch.
 
 ## 2.3 branch
 ### 2.3.0